+++ agent/src/controllers/public/mod.rs
@@ -5,7 +5,7 @@ pub mod nodex_find_identifier;
 pub mod nodex_receive;
 pub mod nodex_verify_didcomm_message;
 pub mod nodex_verify_verifiable_message;
-pub mod send_attributes;
+pub mod send_attribute;
 pub mod send_custom_metric;
 pub mod send_event;
 mod utils;
